Updated parse strategy with test.
This is a copy of the parse_strategy for parsing cuds in oteapi-amiii
with the following changes.

The cuds-file is not downloaded within in the parse strategy.
This is because it is the path specified in the downloadUrl and is therefore
taken care of by the downloadstrategy.

Different methods to get access to the cuds_key. This is copied from
oteapi_dlite.

Updated to have two tests. One that tests the strategy directly and
threfore puts the cuds directly into the session before the parse strategy
is run and one with otelib (which calls the download magic itself)